Move mask to common utilities
