+++ imod/prepare/cleanup.py
@@ -6,7 +6,7 @@ from typing import Optional
 import pandas as pd
 import xarray as xr
 
-from imod.mf6.utilities.mask import mask_arrays
+from imod.common.utilities.mask import mask_arrays
 from imod.prepare.wells import locate_wells, validate_well_columnnames
 from imod.schemata import scalar_None
 from imod.typing import GridDataArray
